Renamed namespace and modified line endings
