--- Assets/General/NodeReflectionEngine/Enumerations/ExposedPropertyTypeEnum.cs
@@ -1,14 +1,0 @@
-namespace NodeReflectionEngine.Enumerations
-{
-    /// <summary>
-    /// Represents the supported types for the exposed properties
-    /// </summary>
-    public enum ExposedPropertyTypeEnum
-    {
-        NONE = 0,
-        BOOL = 1,
-        INT = 2,
-        FLOAT = 3,
-        STRING = 4,
-    }
-}
